+++ data/constants.py
@@ -73,7 +73,7 @@ LOGS_DIR = os.path.join(CONFIG_LOCATION, "logs")
 # Proper usage is "if 'extension'.lower() in ALLOWED_INPUT:"
 JPEG_ALIASES = ["jpg", "jpeg", "jfif", "jif", "jpe"]
 ALLOWED_INPUT_DJXL = ["jxl"]
-ALLOWED_INPUT_CJXL = JPEG_ALIASES + ["png", "gif", "jxl"]
+ALLOWED_INPUT_CJXL = JPEG_ALIASES + ["png", "apng", "gif", "jxl"]
 ALLOWED_INPUT_CJPEGLI = JPEG_ALIASES + ["png", "jxl"]
 ALLOWED_INPUT_IMAGE_MAGICK = JPEG_ALIASES + ["png", "gif", "webp", "jp2", "bmp", "ico", "tiff", "tif"]
 ALLOWED_INPUT_AVIFENC = JPEG_ALIASES + ["png"]
@@ -84,6 +84,7 @@ ALLOWED_RESAMPLING = ("Lanczos", "Point", "Box", "Cubic", "Hermite", "Gaussian",
 
 ALLOWED_INPUT_FILTERS = [
     listToFilter("Supported Images", ALLOWED_INPUT),
+    listToFilter("APNG", ["apng"]),
     listToFilter("AVIF", ["avif"]),
     listToFilter("BMP", ["bmp"]),
     listToFilter("GIF", ["gif"]),