+++ core/worker.py
@@ -277,6 +277,8 @@ class Worker(QRunnable):
                         args.append("-c aom")
                         if self.params["aom_av1_chroma_subsampling"] != "Default":
                             args.append(f"-y {self.params['aom_av1_chroma_subsampling'].replace(':', '')}")
+                        if self.settings["avif_aom_iq_tune"]:  # libaom version >= v3.12.0
+                            args.append("-a tune=iq")
                     case "SVT-AV1-PSY":             # Assuming SVT-AV1 was swapped before compilation
                         args.append("-c svt")
                         args.append("-y 420")       # SVT-AV1 only supports YUV:4:2:0